move the string functions to the library
