+++ lib/lib_string.h
@@ -1,0 +1,68 @@
+/**
+ * @file
+ * String manipulation functions
+ *
+ * @authors
+ * Copyright (C) 2017 Richard Russon <rich@flatcap.org>
+ *
+ * @copyright
+ * This program is free software: you can redistribute it and/or modify it under
+ * the terms of the GNU General Public License as published by the Free Software
+ * Foundation, either version 2 of the License, or (at your option) any later
+ * version.
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
+ * details.
+ *
+ * You should have received a copy of the GNU General Public License along with
+ * this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#ifndef _LIB_STRING_H
+#define _LIB_STRING_H
+
+#include <ctype.h>
+#include <stdio.h>
+
+#define SHORT_STRING 128
+#define STRING       256
+#define LONG_STRING  1024
+#define HUGE_STRING  8192
+
+#define NONULL(x) x ? x : ""
+#define ISSPACE(c) isspace((unsigned char) c)
+#define EMAIL_WSP " \t\r\n"
+
+/* this macro must check for (*c == 0) since isspace(0) has unreliable behavior
+   on some systems */
+#define SKIPWS(c)                                                              \
+  while (*(c) && isspace((unsigned char) *(c)))                                \
+    c++;
+
+int         is_email_wsp(char c);
+int         mutt_atos(const char *str, short *dst);
+int         mutt_atoi(const char *str, int *dst);
+void        mutt_remove_trailing_ws(char *s);
+char *      mutt_skip_whitespace(char *p);
+void        mutt_str_adjust(char **p);
+int         mutt_strcasecmp(const char *a, const char *b);
+const char *mutt_strchrnul(const char *s, char c);
+int         mutt_strcmp(const char *a, const char *b);
+int         mutt_strcoll(const char *a, const char *b);
+const char *mutt_stristr(const char *haystack, const char *needle);
+size_t      mutt_strlen(const char *a);
+char *      mutt_strlower(char *s);
+int         mutt_strncasecmp(const char *a, const char *b, size_t l);
+int         mutt_strncmp(const char *a, const char *b, size_t l);
+void        mutt_str_replace(char **p, const char *s);
+char *      mutt_substrcpy(char *dest, const char *begin, const char *end, size_t destlen);
+char *      mutt_substrdup(const char *begin, const char *end);
+char *      safe_strcat(char *d, size_t l, const char *s);
+char *      safe_strdup(const char *s);
+char *      safe_strncat(char *d, size_t l, const char *s, size_t sl);
+char *      skip_email_wsp(const char *s);
+char *      strfcpy(char *dest, const char *src, size_t dlen);
+
+#endif /* _LIB_STRING_H */
